Fixes #15
- Cleaned code
- New function that checks if user has already taken a certain survey
added to UserBean

+++ src/main/java/com/marketing/bean/ProductBean.java
@@ -18,7 +18,7 @@ public class ProductBean extends AbstractFacade<Product> {
         return getProductByDate(startOfToday);
     }
 
-    public Product getProduct(int id) {
+    public Product getProduct(long id) {
         return find(id);
     }
